Add `cast_shape` spatial query and rename things
